requirements.txt
aqtinstall==3.3.0
beautifulsoup4==4.13.5
Brotli==1.1.0
bs4==0.0.2
certifi==2025.8.3
charset-normalizer==3.4.3
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
humanize==4.13.0
idna==3.10
inflate64==1.0.3
Markdown==3.9
multivolumefile==0.2.3
orjson==3.11.3
patch-ng==1.18.1
protobuf==6.32.1
psutil==7.0.0
py7zr==1.0.0
pybcj==1.0.6
pycryptodomex==3.23.0
pyppmd==1.2.0
PySocks==1.7.1
pyzstd==0.17.0
requests==2.32.5
semantic-version==2.10.0
soupsieve==2.8
texttable==1.7.0
typing_extensions==4.15.0
urllib3==2.5.0
